requirements.txt
aws-cdk-lib==2.86.0
constructs>=10.0.0,<11.0.0
boto3>=1.16.47
boto3-stubs[essential]
boto3-stubs[stepfunctions]
boto3-stubs[sagemaker-a2i-runtime]
boto3-stubs[cloudwatch]
boto3-stubs[logs]
jsii>=1.16.0
DotMap
Pillow
XlsxWriter
pandas
pytz
uuid
